feat: add feature to create order and voucher

+++ src/main/java/iuh/fit/fashionecommercewebsitebackend/repositories/VoucherRepository.java
@@ -3,5 +3,5 @@ package iuh.fit.fashionecommercewebsitebackend.repositories;
 import iuh.fit.fashionecommercewebsitebackend.models.Voucher;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VoucherRepository extends JpaRepository<Voucher, Integer> {
+public interface VoucherRepository extends JpaRepository<Voucher, Long> {
 }
